requirements.txt
streamlit
openai 
msal
selenium
Pillow
requests